# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PromptBridge & Natural Calamity Detector API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * standard | complex — complex adds structured prompts and task breakdown for multi-step work
 */
export type PromptGenerateInputComplexity = typeof PromptGenerateInputComplexity[keyof typeof PromptGenerateInputComplexity];


export const PromptGenerateInputComplexity = {
  standard: 'standard',
  complex: 'complex',
} as const;

export interface PromptGenerateInput {
  /** User's description in Hindi or English */
  description: string;
  /** chatgpt | image_midjourney | image_dalle | image_stable_diffusion | video_ai | writing | coding | business | education | creative | personal */
  taskType: string;
  /** professional | casual | creative | simple | detailed */
  tone?: string;
  targetAudience?: string;
  /** Input language hint (hi=Hindi, en=English) */
  language?: string;
  /** standard | complex — complex adds structured prompts and task breakdown for multi-step work */
  complexity?: PromptGenerateInputComplexity;
}

export interface TaskBreakdownStep {
  step: number;
  title: string;
  description: string;
}

export interface GeneratedPromptVariant {
  prompt: string;
  /** Detailed | Simple | Professional | Creative */
  label: string;
  /** ChatGPT | Midjourney | DALL-E | Stable Diffusion | General */
  platform: string;
  /** @nullable */
  copyUrl?: string | null;
}

export interface PromptGenerateResult {
  prompts: GeneratedPromptVariant[];
  taskType: string;
  /** Explanation of the prompt in Hindi */
  hindiExplanation: string;
  tips?: string[];
  /** Sub-steps for complex tasks (only when complexity=complex) */
  taskBreakdown?: TaskBreakdownStep[];
}

export interface PromptRefineInput {
  originalPrompt: string;
  /** more_detailed | simpler | more_creative | more_professional | different_tone */
  refineAction: string;
  taskType: string;
  additionalContext?: string;
}

export interface SavedPrompt {
  id: number;
  prompt: string;
  taskType: string;
  category: string;
  /** @nullable */
  label?: string | null;
  /** @nullable */
  platform?: string | null;
  /** @nullable */
  originalDescription?: string | null;
  /** @nullable */
  hindiExplanation?: string | null;
  isFavorite: boolean;
  createdAt: string;
}

export interface SavedPromptInput {
  prompt: string;
  taskType: string;
  category: string;
  label?: string;
  platform?: string;
  originalDescription?: string;
  hindiExplanation?: string;
  isFavorite?: boolean;
}

export interface PromptTemplate {
  id: number;
  title: string;
  titleHindi: string;
  description: string;
  descriptionHindi: string;
  prompt: string;
  taskType: string;
  category: string;
  platform: string;
  usageCount: number;
  tags: string[];
}

export interface TaskTypeCount {
  taskType: string;
  count: number;
}

export interface PromptStats {
  totalGenerated: number;
  totalSaved: number;
  totalTemplates: number;
  topTaskTypes: TaskTypeCount[];
  /** Prompts generated in the last 7 days */
  recentActivity: number;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface OpenaiConversationInput {
  title: string;
}

export interface OpenaiMessageInput {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export type OpenaiImageInputSize = typeof OpenaiImageInputSize[keyof typeof OpenaiImageInputSize];


export const OpenaiImageInputSize = {
  '1024x1024': '1024x1024',
  '1536x1024': '1536x1024',
  '1024x1536': '1024x1536',
} as const;

export interface OpenaiImageInput {
  prompt: string;
  size?: OpenaiImageInputSize;
}

export interface OpenaiImageOutput {
  b64_json: string;
}

export interface AgentRunInput {
  /** User's request in Hindi or English */
  description: string;
  taskType: string;
  tone?: string;
  targetAudience?: string;
}

export interface AgentResult {
  agentName: string;
  /** Short role description shown to user */
  agentRole: string;
  agentEmoji: string;
  /** The AI model that powered this response (e.g. gpt-5.4, claude-sonnet-4-6) */
  model?: string;
  /** The actual AI-generated response/content */
  content: string;
  /** text | markdown | code | image_base64 | prompt_only */
  contentType: string;
  language?: string;
  /** @nullable */
  hindiSummary?: string | null;
  /**
     * Base64-encoded image for image generation results
     * @nullable
     */
  imageBase64?: string | null;
  /**
     * URL to open the AI platform with the generated prompt pre-filled
     * @nullable
     */
  navigateTo?: string | null;
}

export interface AgentAlternative {
  name: string;
  emoji: string;
  url: string;
  why: string;
}

export interface AgentRecommendation {
  /** Name of the best AI for this task */
  bestAI: string;
  bestAIEmoji: string;
  /** URL to the best AI platform */
  bestAIUrl: string;
  /** Whether PromptBridge can give a direct response for this task */
  canDirectRespond: boolean;
  /** Why this AI is best (English) */
  reason: string;
  /** Why this AI is best (Hindi) */
  reasonHindi: string;
  alternatives?: AgentAlternative[];
}

export interface AgentRunResult {
  agents: AgentResult[];
  taskType: string;
  /** Brief Hindi overview of all responses */
  hindiOverview: string;
  recommendation?: AgentRecommendation;
}

export interface OpenaiError {
  error: string;
}

export type ListSavedPromptsParams = {
category?: string;
taskType?: string;
};

export type ListPromptTemplatesParams = {
category?: string;
taskType?: string;
};