# lib/api-zod/src/generated/types/openaiError.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PromptBridge & Natural Calamity Detector API
 * OpenAPI spec version: 0.1.0
 */

export interface OpenaiError {
  error: string;
}