# lib/api-zod/src/generated/types/openaiImageOutput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PromptBridge & Natural Calamity Detector API
 * OpenAPI spec version: 0.1.0
 */

export interface OpenaiImageOutput {
  b64_json: string;
}